move old integration tests into raftstore
