--- test/basic_test.go
@@ -1,9 +1,0 @@
-package test
-
-import (
-	"time"
-)
-
-var (
-	testWaitTimeout = time.Minute
-)
